% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unzip_bin.character}
\alias{unzip_bin.character}
\title{Unzip file to raw vectors}
\usage{
\method{unzip_bin}{character}(x, ...)
}
\arguments{
\item{x}{Name of zip file.}

\item{...}{Ignored parameters.}
}
\description{
Unzip file to raw vectors
}
