% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unzip_bin.raw}
\alias{unzip_bin.raw}
\title{Raw unzip to raw vectors}
\usage{
\method{unzip_bin}{raw}(x, ...)
}
\arguments{
\item{x}{Raw zip vector to unzip.}

\item{...}{Ignored parameters.}
}
\value{
List of member raw vectors.
}
\description{
Raw unzip to raw vectors
}
