% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unzip_bin}
\alias{unzip_bin}
\title{Unzip to raw vectors}
\usage{
unzip_bin(x, ...)
}
\arguments{
\item{x}{Name of zip file or raw zip vector.}

\item{...}{Passed to method.}
}
\description{
Unzip to raw vectors
}
